:building_construction: move front-end register code
